requirements.txt
addict
clip
colour
einops
fvcore
h5py
json_tricks
loguru
matplotlib
natsort
numpy==1.23.5
open3d
opencv-contrib-python
psutil
PyYAML
scipy
setuptools
t2m
timm
tqdm
transforms3d
yacs
smplx
torchgeometry
